+++ CHANGELOG.md
@@ -3,6 +3,7 @@ https://keepachangelog.com/en/1.0.0/
 
 ## [1.21.4] - 2023-01-29
 - `StringConcatenatedInLoop`: Concatenations inside object creation expressions no longer trigger a warning
+- `StringConcatenatedInLoop`: Does not trigger when assigning to the loop variable
 
 ## [1.21.3] - 2023-01-25
 - `LoopedRandomInstantiation`: No longer triggers when a seed is passed to the constructor